requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.42.0
sqlalchemy==2.0.48
psycopg2-binary==2.9.11
python-multipart==0.0.22
httpx==0.28.1
openpyxl==3.1.5
jinja2==3.1.6
# Email Mining v2 hardening
filetype==1.2.0
charset-normalizer==3.4.6
# Credential encryption
cryptography==46.0.5
# Observability
loguru==0.7.3
prometheus-fastapi-instrumentator==7.1.0
sentry-sdk[fastapi]==2.55.0
# Rate limiting
slowapi==0.1.9
# Fuzzy matching
rapidfuzz>=3.0.0
# PDF generation
weasyprint==68.1
# Scheduling
apscheduler==3.11.2,<4.0
# Migrations
alembic==1.18.4
# Caching
redis[hiredis]==7.3.0
orjson>=3.9.0
# CSRF protection
starlette-csrf==3.0.0
# NetComponents integration
beautifulsoup4==4.14.3
patchright==1.58.2
# Typed settings from env vars (fail-fast validation at startup)
pydantic-settings==2.13.1
# SSE streaming for HTMX sourcing progress
sse-starlette>=1.6.0
# Anthropic Claude API (used by claude_client, enrichment, batch extraction)
anthropic>=0.40.0
# HTML sanitization for untrusted email content (XSS prevention)
nh3>=0.2.0
# Safe XML parsing (XXE prevention)
defusedxml>=0.7.0
